chore: fix sonarqube duplications
