+++ code/packages/sdk/src/plugins/nodes-distance-snapping/nodes-distance-snapping.ts
@@ -24,9 +24,8 @@ import {
   WEAVE_NODES_DISTANCE_SNAPPING_PLUGIN_KEY,
 } from './constants';
 import type { KonvaEventObject } from 'konva/lib/Node';
-import type { WeaveNodesSelectionPlugin } from '../nodes-selection/nodes-selection';
 import type { BoundingBox } from '@inditextech/weave-types';
-import { getTargetAndSkipNodes, getVisibleNodesInViewport } from '@/utils';
+import { getTargetAndSkipNodes, getVisibleNodes } from '@/utils';
 import type { Context } from 'konva/lib/Context';
 
 export class WeaveNodesDistanceSnappingPlugin extends WeavePlugin {
@@ -108,9 +107,16 @@ export class WeaveNodesDistanceSnappingPlugin extends WeavePlugin {
       return;
     }
 
+    const stage = this.instance.getStage();
     this.referenceLayer = nodeParent as unknown as Konva.Layer | Konva.Group;
 
-    const visibleNodes = this.getVisibleNodes(nodeParent, skipNodes);
+    const visibleNodes = getVisibleNodes(
+      this.instance,
+      stage,
+      nodeParent,
+      skipNodes,
+      this.referenceLayer
+    );
     // find horizontally intersecting nodes
     const {
       intersectedNodes: sortedHorizontalIntersectedNodes,
@@ -719,70 +725,6 @@ export class WeaveNodesDistanceSnappingPlugin extends WeavePlugin {
     }
 
     return { intersectedNodes, intersectedNodesWithDistances };
-  }
-
-  private getVisibleNodes(nodeParent: Konva.Node, skipNodes: string[]) {
-    const stage = this.instance.getStage();
-
-    const nodesSelection =
-      this.instance.getPlugin<WeaveNodesSelectionPlugin>('nodesSelection');
-
-    if (nodesSelection) {
-      nodesSelection.getTransformer().hide();
-    }
-
-    const nodes = getVisibleNodesInViewport(stage, this.referenceLayer);
-
-    const finalVisibleNodes: Konva.Node[] = [];
-
-    // and we snap over edges and center of each object on the canvas
-    nodes.forEach((node) => {
-      const actualNodeParent = this.instance.getNodeContainer(node);
-
-      if (actualNodeParent?.getAttrs().id !== nodeParent?.getAttrs().id) {
-        return;
-      }
-
-      if (node.getParent()?.getAttrs().nodeType === 'group') {
-        return;
-      }
-
-      if (skipNodes.includes(node.getParent()?.getAttrs().nodeId)) {
-        return;
-      }
-
-      if (skipNodes.includes(node.getAttrs().id ?? '')) {
-        return;
-      }
-
-      if (
-        node.getParent() !== this.referenceLayer &&
-        !node.getParent()?.getAttrs().nodeId
-      ) {
-        return;
-      }
-
-      if (
-        node.getParent() !== this.referenceLayer &&
-        node.getParent()?.getAttrs().nodeId
-      ) {
-        const realNode = stage.findOne(
-          `#${node.getParent()?.getAttrs().nodeId}`
-        ) as Konva.Group;
-
-        if (realNode && realNode !== this.referenceLayer) {
-          return;
-        }
-      }
-
-      finalVisibleNodes.push(node);
-    });
-
-    if (nodesSelection) {
-      nodesSelection.getTransformer().show();
-    }
-
-    return finalVisibleNodes;
   }
 
   private drawSizeGuidesHorizontally(
